+++ pet-nutrition-service/nutrition-service-update.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 
 if [ ! -z "$1" ]
 then
@@ -9,12 +10,17 @@ else
       export REGION="us-east-1"
 fi
 
-export ACCOUNT_ID=`aws sts get-caller-identity | jq .Account -r`
-export ECR_URL=${ACCOUNT_ID}.dkr.ecr.${REGION}.amazonaws.com
-aws ecr get-login-password --region ${REGION} | docker login --username AWS --password-stdin ${ECR_URL}
+export ACCOUNT_ID=$(aws sts get-caller-identity | jq .Account -r)
+if [ -z "$ACCOUNT_ID" ]; then
+    echo "Error: Failed to get AWS account ID"
+    exit 1
+fi
+
+export ECR_URL="${ACCOUNT_ID}.dkr.ecr.${REGION}.amazonaws.com"
+aws ecr get-login-password --region "${REGION}" | docker login --username AWS --password-stdin "${ECR_URL}"
 
 docker build -t nutrition-service . --no-cache
-docker tag nutrition-service:latest ${ECR_URL}/nodejs-petclinic-nutrition-service:latest
-docker push ${ECR_URL}/nodejs-petclinic-nutrition-service:latest
+docker tag nutrition-service:latest "${ECR_URL}/nodejs-petclinic-nutrition-service:latest"
+docker push "${ECR_URL}/nodejs-petclinic-nutrition-service:latest"
 
 kubectl delete pods -l io.kompose.service=nutrition-service-nodejs
